requirements.txt
asgiref==3.5.2
async-generator==1.10
attrs==21.4.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.4
distlib==0.3.4
Django==4.0.5
django-ckeditor==6.4.2
django-js-asset==2.0.0
django-phonenumber-field==6.3.0
django-webpush==0.3.4
filelock==3.7.1
h11==0.13.0
http-ece==1.1.0
idna==3.3
outcome==1.2.0
phonenumbers==8.12.51
Pillow==9.2.0
platformdirs==2.5.2
py-vapid==1.8.2
pycparser==2.21
pyOpenSSL==22.0.0
PySocks==1.7.1
pytz==2020.1
pywebpush==1.9.4
requests==2.28.1
selenium==4.3.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
stripe==3.5.0
trio==0.21.0
trio-websocket==0.9.2
tzdata==2022.1
urllib3==1.26.10
virtualenv==20.14.1
wsproto==1.1.0
